Move player and enemy battle info to their own files
