+++ src/ui/battle-info/enemy-battle-info.ts
@@ -1,0 +1,9 @@
+import BattleInfo from "./battle-info";
+
+export class EnemyBattleInfo extends BattleInfo {
+  constructor() {
+    super(140, -141, false);
+  }
+
+  setMini(_mini: boolean): void {} // Always mini
+}
